implement actionbar

+++ app/src/main/java/org/matt/calculatorapp/Presenter/AppPresenter.java
@@ -28,7 +28,7 @@ public class AppPresenter implements Presenter {
     }
 
     @Override
-    public int getResult() {
+    public double getResult() {
         return calculator.getResult();
     }
